[LLT-5319] Move CryptoStepRng to telio-utils
